// src/fichas/enums/ficha.enums.ts
export enum GradoInstruccion {
  ILETRADO = 'Iletrado',
  PRIMARIA = 'Primaria',
  SECUNDARIA = 'Secundaria',
  SUPERIOR_TECNICO = 'Superior Técnico',
  SUPERIOR_UNIVERSITARIO = 'Superior Universitario',
}
export enum EstadoCivil {
  SOLTERO = 'Soltero',
  CONVIVIENTE_CASADO = 'Conviviente/Casado',
  VIUDO = 'Viudo',
  DIVORCIADO_SEPARADO = 'Divorciado/Separado',
}
export enum Aseguramiento {
  SI = 'Sí',
  NO = 'No',
  ES_SALUD = 'EsSalud',
  OTRO = 'Otro',
}
export enum CondicionOcupacional {
  PROFESIONAL_INDEPENDIENTE = 'Profesional/Independiente',
  PROFESIONAL_DEPENDIENTE = 'Profesional/Dependiente',
  TECNICO_MANDO_MEDIO = 'Técnico Mando Medio',
  TRABAJADOR_INFORMAL = 'Trabajador Informal',
  EVENTUAL = 'Eventual',
  PENSIONISTA = 'Pensionista',
  SU_CASA = 'Su casa',
  PERMANENTE = 'Permanente',
  SIN_OCUPACION = 'Sin ocupación',
}
export enum GradoDependenciaEconomica {
  HASTA_3 = 'Hasta 3 miembros',
  MAS_DE_3 = 'Más de 3 miembros',
}
export enum TramoIngreso {
  NINGUNA = 'Ninguna',
  MENOS_1_SMV = 'Menos de 1 SMV',
  DE_1_A_2_SMV = 'De 1 a 2 SMV',
  DE_2_A_3_SMV = 'De 2 a 3 SMV',
  DE_3_A_4_SMV = 'De 3 a 4 SMV',
  MAS_DE_SMV = 'Más de SMV',
}
export enum Tenencia {
  PROPIA = 'Propia',
  ALQUILADA = 'Alquilada',
  INVASION = 'Invasión',
  GUARDIANIA = 'Guardianía',
  ALOJADO = 'Alojado',
}
export enum MaterialConstruccion {
  NOBLE_ACABADO = 'Noble/Acabado',
  NOBLE_SIN_ACABAR = 'Noble sin acabar',
  MIXTO = 'Mixto',
  RUSTICO = 'Rústico',
  PRECARIO = 'Precario',
}
export enum ServiciosBasicos {
  COMPLETO = 'Completo',
  PARCIAL = 'Parcial',
  SIN_SERVICIOS = 'Sin servicios básicos',
}
export enum EquipamientoHogar {
  NO_CUENTA = 'No cuenta con artefactos',
  DE_1_A_2 = '1 a 2',
  DE_3_A_MAS = '3 a más',
}
export enum FactorRiesgo {
  NINOS_DESNUTRIDOS = 'Niños Desnutridos',
  TUBERCULOSIS = 'Tuberculosis',
  ETS_SIDA = 'ETS-Sida',
  INCAPACIDAD = 'Incapacidad Física o Mental',
  ANTECEDENTES_PENALES = 'Antecedentes Penales',
  FARMACO_DEPENDENCIA = 'Fármaco Dependencia',
  ABANDONO_FAMILIAR = 'Abandono Familiar',
  VIOLENCIA_FAMILIAR = 'Violencia Familiar',
  PROSTITUCION = 'Prostitución',
}
export enum Categoria {
  A = 'A',
  B = 'B',
  C = 'C',
  Z = 'Z',
}